Fixed create item feature
