+++ .devcontainer/docker-compose.yml
@@ -13,10 +13,10 @@ services:
       - ..:/app
     command: sleep infinity
     environment:
-      PORT: 8000
+      PORT: 8080
       FLASK_APP: wsgi:app
       FLASK_DEBUG: "True"
-      GUNICORN_BIND: "0.0.0.0:8000"
+      GUNICORN_BIND: "0.0.0.0:8080"
       DATABASE_URI: postgresql+psycopg://postgres:pgs3cr3t@postgres:5432/postgres
     networks:
       - dev
