+++ features/environment.py
@@ -6,7 +6,7 @@ from os import getenv
 from selenium import webdriver
 
 WAIT_SECONDS = int(getenv("WAIT_SECONDS", "60"))
-BASE_URL = getenv("BASE_URL", "http://localhost:8000")
+BASE_URL = getenv("BASE_URL", "http://localhost:8080")
 DRIVER = getenv("DRIVER", "chrome").lower()
 
 
@@ -14,6 +14,7 @@ def before_all(context):
     """Executed once before all tests"""
     context.base_url = BASE_URL
     context.wait_seconds = WAIT_SECONDS
+    context.scenario_data = {}
     # Select either Chrome or Firefox
     if "firefox" in DRIVER:
         context.driver = get_firefox()
