+++ wsgi.py
@@ -5,7 +5,7 @@ Web Server Gateway Interface (WSGI) entry point
 import os
 from service import create_app
 
-PORT = int(os.getenv("PORT", "8000"))
+PORT = int(os.getenv("PORT", "8080"))
 
 app = create_app()
